refactor: move PressableSourceProps
